pyproject.toml
[project]
name = "agenticaiworkspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "faiss-cpu>=1.13.2",
    "ipython>=9.9.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-experimental",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.7",
    "langgraph-cli[inem,inmem]>=0.4.14",
    "langgraph-supervisor>=0.0.31",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.3.0",
    "wikipedia>=1.4.0",
]
